// src/shared/constants/ROUTES_PATHS.d.ts
export declare const ROUTES_PATHS: {
    home: {
        url: string;
        title: string;
        i18nKey: string;
    };
    services: {
        url: string;
        title: string;
        i18nKey: string;
    };
    price: {
        url: string;
        title: string;
        i18nKey: string;
    };
    machineries: {
        url: string;
        title: string;
        i18nKey: string;
    };
    equipments: {
        url: string;
        title: string;
        i18nKey: string;
    };
    projects: {
        url: string;
        title: string;
        i18nKey: string;
    };
    reviews: {
        url: string;
        title: string;
        i18nKey: string;
    };
    faq: {
        url: string;
        title: string;
        i18nKey: string;
    };
};